requirements/pip.txt
django-decouple==2.1
dj-database-url==0.5.0
Django==4.2.4
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
django-rest-passwordreset==1.3.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg==1.21.7
